:sparkles: refactor create and update talk
