+++ src/components/form/validators.js
@@ -1,4 +1,3 @@
 /* eslint-disable import/prefer-default-export */
-import capitalize from 'lodash/capitalize'
 
-export const required = (value, field, form, name) => (value ? undefined : `${capitalize(name)} is required`)
+export const required = value => (value ? undefined : 'Required')
