+++ src/screens/speaker/speaker.jsx
@@ -8,8 +8,7 @@ import Contributors from 'screens/components/contributors'
 import Profile from 'screens/components/profile'
 import Sidebar from './sidebar'
 import MyTalks from './myTalks'
-import TalkCreate from './talk/create'
-import TalkEdit from './talk/edit'
+import { TalkEdit, TalkCreate } from './talk/form'
 import TalkSubmission from './talk/submission'
 import Talk from './talk/page'
 import EventPage from './event/page'
