--- src/screens/speaker/talk/components/talkForm/index.js
@@ -1,1 +1,0 @@
-export { default } from './talkForm'
