--- src/screens/speaker/talk/create/index.js
@@ -1,1 +1,0 @@
-export { default } from './talkCreate.container'
