--- src/screens/speaker/talk/edit/index.js
@@ -1,1 +1,0 @@
-export { default } from './talkEdit.container.js'
