+++ src/screens/speaker/talk/form/index.js
@@ -1,0 +1,2 @@
+export { default as TalkEdit } from './talkEdit.container'
+export { default as TalkCreate } from './talkCreate.container'
